rename extensions to helpers
