+++ tests/functional/test_routes.py
@@ -4,7 +4,7 @@ from flask_login import current_user
 
 from responses import RequestsMock, GET
 from werkzeug.security import generate_password_hash
-from archivy.extensions import get_max_id, get_db
+from archivy.helpers import get_max_id, get_db
 
 
 def test_get_index(test_app, client: FlaskClient):
